refactoring  the code
