+++ package.json
@@ -64,6 +64,7 @@
     "vue-loader": "^16.5.0",
     "vue-router": "^4.0.0-0",
     "vuex": "^4.0.0-0",
+    "vuex-persist": "^3.1.3",
     "vuex-persistedstate": "^4.0.0"
   },
   "proxy": "http://localhost:3000/",
