+++ src/templates/pages/cart.pug
@@ -5,7 +5,6 @@ block prepend head
 block title
   +page-header(pageName)
 block contentBody
-  .cells
-    .cell.cell-xl-8
-      #app
+  #app
+
